refactor: 로또 구매에 관한 기능 클래스 분리
